added addition of estates

+++ Estate/Controllers/EstateController.cs
@@ -1,6 +1,9 @@
-﻿using System.Linq;
+﻿using System.IO;
+using System.Linq;
+using System.Threading.Tasks;
 using Estate.Data;
 using Estate.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Estate.Controllers
@@ -32,9 +35,18 @@ namespace Estate.Controllers
             return View(model);
         }
 
-        [HttpPost]       
+        [HttpPost]
         public ActionResult Delete(int id)
         {
+            var appartment = context.Appartments.Find(id);
+
+            if (!System.IO.File.Exists("wwwroot\\images\\" + appartment.Images))
+            {
+                return RedirectToAction("Manage");
+            }
+
+            System.IO.File.Delete("wwwroot\\images\\" + appartment.Images);
+
             context.Appartments.Remove(context.Appartments.Find(id));
             context.SaveChanges();
             return RedirectToAction("Manage");
@@ -43,6 +55,31 @@ namespace Estate.Controllers
         public ActionResult Add()
         {
             return View();
+        }
+
+        [HttpPost]
+        public async Task<ActionResult> Submit(IFormFile file, string addres, int price, string description)
+        {
+            if (!file.ContentType.Contains("image"))
+            {
+                return View("Add");
+            }
+
+            var fileName = "wwwroot\\images\\"
+                + file.FileName.Split(new char[] { '\\' }).Last();
+
+            using (var stream = new FileStream(fileName, FileMode.Create))
+            {
+                await file.CopyToAsync(stream);
+            }
+            fileName = fileName.Split(new char[] { '\\' }).Last();
+            Appartment add = new Appartment(0, addres, price,
+                fileName, HttpContext.User.Identity.Name,
+                description);
+            context.Appartments.Add(add);
+            context.SaveChanges();
+
+            return RedirectToAction("Add");
         }
     }
 }
